requirements.txt
appdirs==1.4.4
distlib==0.3.1
django-restframework==0.0.1
filelock==3.0.12
six==1.15.0
virtualenv==20.0.31
virtualenvwrapper-win==1.2.6
